MISC - fix form_helper region_select
