+++ lib/china_regions.rb
@@ -1,5 +1,5 @@
 require "china_regions/version"
 require 'china_regions/engine' if defined? Rails
-require "china_regions/action_view/extension"
+require "china_regions/helpers/form_helper"
 
 module ChinaRegions;end
