+++ lib/china_regions/engine.rb
@@ -6,7 +6,7 @@ module ChinaRegions
   else
     class Railtie < ::Rails::Railtie
       ActiveSupport.on_load(:action_view) do
-        ::ActionView::Base.send :include, Penmanship::ActionView::Extension
+        ::ActionView::Base.send :include, ChinaRegions::Helpers::FormHelper
       end  
     end
   end
